pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tv-indicator-research"
version = "0.1.0"
description = "TradingView indicator intake, analysis, translation, and backtesting pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "ccxt>=4.4.0",
  "numpy>=1.26",
  "pandas>=2.2",
  "PyYAML>=6.0",
  "vectorbt>=0.26",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.9.0",
]

[project.scripts]
tvir = "tv_indicators.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
